Commit final, rumo ao 20

+++ RIUPP/RIUPP/Controllers/AreasController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -18,24 +19,45 @@ namespace RIUPP.Controllers
         /// </summary>
         private readonly RIUPPDB _context;
 
+        /// <summary>
+        /// Variável que identifica, dentro da base de dados a parte do utilizador.
+        /// </summary>
+        private readonly UserManager<IdentityUser> _userManager;
 
         // Construtor
-        public AreasController(RIUPPDB context)
-        {
+        public AreasController(RIUPPDB context, UserManager<IdentityUser> userManager){
             _context = context;
+            _userManager = userManager;
         }
 
-        // Retorna a vista com todas as Áreas
-        // GET: Areas
-        public async Task<IActionResult> Index()
-        {
+
+        /// <summary>
+        /// Retorna a vista com todas as Áreas, GET: Areas
+        /// </summary>
+        public async Task<IActionResult> Index(){
+            if (User.Identity.IsAuthenticated){
+                //verifica se o user tem a conta suspensa
+                var user = await _userManager.GetUserAsync(User);
+                var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                if (util.Suspenso) return View("Suspenso");
+            }
             return View(await _context.Areas.ToListAsync());
         }
 
-        // Mostra os detalhes da Área em questão
-        // GET: Areas/Details/5
+
+        /// <summary>
+        /// Mostra os detalhes da Área em questão, GET: Areas/Details/5
+        /// </summary>
         public async Task<IActionResult> Details(int? id)
                 {
+                    if (User.Identity.IsAuthenticated)
+                    {
+                        //verifica se o user tem a conta suspensa
+                        var user = await _userManager.GetUserAsync(User);
+                        var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                        if (util.Suspenso) return View("Suspenso");
+                    }
+
                     if (id == null)
                     {
                         return NotFound();
@@ -57,20 +79,32 @@ namespace RIUPP.Controllers
                 }
         
                 [Authorize(Roles = "Gestor")]
-                // GET: Areas/Create 
-                public IActionResult Create()
-                {
+                /// <summary>
+                /// GET: Areas/Create 
+                /// </summary>
+                public async Task<IActionResult> CreateAsync(){
+                    //verifica se o user tem a conta suspensa
+                    var user = await _userManager.GetUserAsync(User);
+                    var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                    if (util.Suspenso) return View("Suspenso");
+
                     return View();
                 }
+                
 
-                // POST: Areas/Create
-                // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-                // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+                /// <summary>
+                /// POST: Areas/Create
+                /// To protect from overposting attacks, enable the specific properties you want to bind to, for more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+                /// </summary>
                 [HttpPost]
                 [ValidateAntiForgeryToken]
                 [Authorize(Roles = "Gestor")]
-                public async Task<IActionResult> Create([Bind("Id,Nome,Designacao")] Area area)
-                {
+                public async Task<IActionResult> Create([Bind("Id,Nome,Designacao")] Area area){
+                    //verifica se o user tem a conta suspensa
+                    var user = await _userManager.GetUserAsync(User);
+                    var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                    if (util.Suspenso) return View("Suspenso");
+
                     if (ModelState.IsValid)
                     {
                         _context.Add(area);
@@ -80,9 +114,16 @@ namespace RIUPP.Controllers
                     return View(area);
                 }
 
-                // GET: Areas/Edit/5
+                /// <summary>
+                /// GET: Areas/Edit/5
+                /// </summary>
                 [Authorize(Roles = "Gestor")]
                 public async Task<IActionResult> Edit(int? id){
+                    //verifica se o user tem a conta suspensa
+                    var user = await _userManager.GetUserAsync(User);
+                    var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                    if (util.Suspenso) return View("Suspenso");
+
                     if (id == null)
                     {
                         return NotFound();
@@ -96,14 +137,19 @@ namespace RIUPP.Controllers
                     return View(area);
                 }
 
-                
-                // POST: Areas/Edit/5
-                // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-                // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+                /// <summary>
+                /// POST: Areas/Edit/5
+                /// To protect from overposting attacks, enable the specific properties you want to bind to, for more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+                /// </summary>
                 [HttpPost]
                 [ValidateAntiForgeryToken]
                 [Authorize(Roles = "Gestor")]
                 public async Task<IActionResult> Edit(int id, [Bind("Id,Nome,Designacao")] Area area){
+                    //verifica se o user tem a conta suspensa
+                    var user = await _userManager.GetUserAsync(User);
+                    var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                    if (util.Suspenso) return View("Suspenso");
+
                     if (id != area.Id){
                         return NotFound();
                     }
@@ -128,9 +174,16 @@ namespace RIUPP.Controllers
                     return View(area);
                 }
 
-                // GET: Areas/Delete/5
+                /// <summary>
+                /// GET: Areas/Delete/5
+                /// </summary>
                 [Authorize(Roles = "Gestor")]
                 public async Task<IActionResult> Delete(int? id){
+                    //verifica se o user tem a conta suspensa
+                    var user = await _userManager.GetUserAsync(User);
+                    var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                    if (util.Suspenso) return View("Suspenso");
+
                     if (id == null)
                     {
                         return NotFound();
@@ -146,12 +199,19 @@ namespace RIUPP.Controllers
                     return View(area);
                 }
 
-                // POST: Areas/Delete/5
+                /// <summary>
+                /// POST: Areas/Delete/5
+                /// </summary>
                 [HttpPost, ActionName("Delete")]
                 [ValidateAntiForgeryToken]
                 [Authorize(Roles = "Gestor")]
                 public async Task<IActionResult> DeleteConfirmed(int id){
-                    var area = await _context.Areas.FindAsync(id);
+                //verifica se o user tem a conta suspensa
+                var user = await _userManager.GetUserAsync(User);
+                var util = await _context.Utilizadores.FirstOrDefaultAsync(u => u.Aut == user.Id);
+                if (util.Suspenso) return View("Suspenso");
+
+                var area = await _context.Areas.FindAsync(id);
                     _context.Areas.Remove(area);
                     await _context.SaveChangesAsync();
                     return RedirectToAction(nameof(Index));
